requirements.txt
# requirements.txt
streamlit
requests
websocket-client
pillow
python-dotenv
psutil
gputil
safetensors
torch
torchvision
